任务5

+++ demo.js
@@ -2,14 +2,16 @@
 
 var fs = require("fs")
 var dirName = process.argv[2]
+if(fs.existsSync(dirName)){
+    console.log("error:exists" + " "+dirName);
+    process.exit(1);
+}else{
+ fs.mkdirSync(dirName);
+ process.chdir(dirName);
+ fs.mkdirSync("css");
+ fs.mkdirSync("js");
 
-fs.mkdirSync(dirName)
-process.chdir(dirName)
-fs.mkdirSync("css")
-fs.mkdirSync("js")
-
-fs.writeFileSync("index.html","<DOCTYPE><title>Hello</title><h1>Hi</h1>")
-fs.writeFileSync("css/style.css","h1{color: red;}")
-fs.writeFileSync("js/main.js","var string = \"Hello World\" alert(string)")
-
-process.exit(0)
+ fs.writeFileSync("index.html","<DOCTYPE><title>Hello</title><h1>Hi</h1>");
+ fs.writeFileSync("css/style.css","h1{color: red;}");
+ fs.writeFileSync("js/main.js","var string = \"Hello World\" alert(string)");
+ process.exit(0);}
